+++ libsolidity/analysis/ControlFlowBuilder.cpp
@@ -290,7 +290,7 @@ bool ControlFlowBuilder::visit(ModifierInvocation const& _modifierInvocation)
 			appendControlFlow(*argument);
 
 	auto modifierDefinition = dynamic_cast<ModifierDefinition const*>(
-		_modifierInvocation.name()->annotation().referencedDeclaration
+		_modifierInvocation.name().annotation().referencedDeclaration
 	);
 	if (!modifierDefinition) return false;
 	solAssert(!!modifierDefinition, "");