+++ libsolidity/analysis/PostTypeChecker.cpp
@@ -217,9 +217,9 @@ struct OverrideSpecifierChecker: public PostTypeChecker::Checker
 
 	void endVisit(OverrideSpecifier const& _overrideSpecifier) override
 	{
-		for (ASTPointer<UserDefinedTypeName> const& override: _overrideSpecifier.overrides())
+		for (ASTPointer<IdentifierPath> const& override: _overrideSpecifier.overrides())
 		{
-			Declaration const* decl  = override->annotation().referencedDeclaration;
+			Declaration const* decl = override->annotation().referencedDeclaration;
 			solAssert(decl, "Expected declaration to be resolved.");
 
 			if (dynamic_cast<ContractDefinition const*>(decl))